+++ src/app/(blog)/(content)/memos/[slug]/memo-detail-page-content.tsx
@@ -1,7 +1,7 @@
 import { AdminEditLinkClient } from "@/components/admin/admin-links.client";
 import { renderMDX } from "@/components/mdx/mdx-content";
 import { TableOfContents } from "@/components/table-of-contents.client";
-import type { Memo } from "@/libs/contents/types";
+import type { Memo } from "@/libs/contents/types/legacy";
 import { cn } from "@/utils/cn";
 import { MemoBackLink } from "./memo-back-link";
