+++ src/app/sitemap.ts
@@ -1,7 +1,7 @@
 import type { MetadataRoute } from "next";
 import { sanitizeSlug } from "@/keystatic/libs/slug";
-import { getMemoList } from "@/libs/contents/memo";
-import { getPostList } from "@/libs/contents/post";
+import { getMemoList } from "@/libs/contents/services/memo";
+import { getPostList } from "@/libs/contents/services/post";
 
 // TODO : 추후 updatedAt을 추가 후 수정
 export default async function sitemap(): Promise<MetadataRoute.Sitemap> {
