+++ src/libs/contents/types/legacy.ts
